Move atlas files
